Refactor vector_util interface. Add unit tests
